refactor: refactor static methods
